+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - `hide_inactive_statusline` option deprecated from `lualine` plugin
+- `javascriptTSTagAttribute` highlight removed
+- `TSTagAttribute` highlight added
 
 ## [v0.0.3]- 11 Dec 2021